Crate dxgi

FFI bindings to DXGI

Reexports

pub use constants::*;

pub use enumerations::*;

pub use structures::*;

pub use interfaces::*;

Modules

constants
Constants and related typedefs provided by DXGI

enumerations
Enumerations provided by DXGI.

interfaces
Interfaces provided by DXGI.

structures
Structures provided by DXGI.

Macros

c_mtdcall!
Call a method of a C++ object represented as a C struct

c_vtable!
Macro used to make making bindings for C representations of C++ classes more convenient.

c_vtable_main!

c_vtable_methods_to_trait!

c_vtable_pre!
Handle the pub keyword for vtable and trait, and "with heirs" phrase before main expansion

c_vtable_struct!

define_guid!
Rust equivalent to windows C DEFINE_GUID macro

impl_c_vtable_trait!

Functions

CreateDXGIFactory1